Added a new property for the player threshold. If there are not enough players on the server to meet the threshold then nobody will be kicked. Added a command to allow setting it as well (playerCount). Note that this comparison is a less than (players online < threshold, don't kick). If you have the same number of players online as the threshold then it will kick as normal.
Also added a timer to the playersToKick mechanism to make sure that we're not hosed if we fail to clear out the collection on one of the kick tasks. It'll continue to warn you about failures every kick check time for a minute (currently hardcoded) at which point you should get a log message about failing to kick players and the collection will be cleared, resetting the mechanism to ready again.

Bumped the version number to 0.5. Also added the version number to the settings file.