Digest of the messages given in my personal subversion repository:
Today
Removed the EventListener because it could not be used for events as a generic listener as previously hoped.  Apparently Bukkit tries to cast it to a PlayerListener or BlockListener when it sends the events.  Right now there's a somewhat kludgy method of detecting whether it's a block event or not by checking if it starts with "BLOCK."  Everything else is assumed to be a player event.

Added a command for the block items for idlers ability.

Added ability to listen to PLAYER_COMMAND_PREPROCESS event.  This will need to be explicitly added to the list now.

Removed all the old event listening code.

Finally fixed the annoying bug that required a server restart in order to properly change the kick timeout time.  As a result of this, all local variables in the main class (AfkBooter) were removed and instead directly queried from the Settings class.  This should remove additional points of failure/confusion.

5/6/11
Created a new generic listener object that should listen for both the block and player events.  Adapted the EventCatalog to make use of the new methodology as provided by the Settings class.

Also commented out the old player and block listener registrations in the main plugin class; replaced them with registrations for the new generic listener.

Note that this needs to be tested.  It should be fine without having implemented all of the player and block events as empty.  Bukkit should only be sending events to it for the ones that were registered for.

5/4/11
Added initial settings handling for putting in a list of events to listen to.  Still need to flesh out the event catalog and make it able to parse these events from their split names.

4/26/11
Added initial capability to block player item pickups when a player is marked as idle.

Added checking for event cancellation in event listeners, where available.  I had initially thought that this was unnecessary for a plugin listening on Monitor, but I am no longer so sure.

Will probably also want to revamp the event registration system for what constitutes activity to the listing system that was previously planned; the current one is painfully inextensible.